Fix typos and improve wording in gpt_guidance.txt
The commit rectifies typographical errors and refines some phrasing to ensure clearer instructions in gpt_guidance.txt. Specifically, it corrects missing words in sentences about UNIT TESTS and the order of steps. In addition, the documentation about commit action after a refactor is now more inclusive and provides an alternative if the individual can't perform the commit.